+++ message/type_test.go
@@ -4,8 +4,10 @@ import (
 	"reflect"
 
 	. "github.com/dogmatiq/configkit/message"
+	"github.com/dogmatiq/dogma"
 	. "github.com/dogmatiq/enginekit/enginetest/stubs"
 	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
 )
 
@@ -157,9 +159,20 @@ var _ = Describe("type Type", func() {
 	})
 
 	Describe("func String()", func() {
-		It("returns the string representation of the type", func() {
-			t := TypeOf(CommandA1)
-			Expect(t.String()).To(Equal("stubs.CommandStub[TypeA]"))
-		})
+		type NonGenericCommand struct {
+			CommandStub[TypeA]
+		}
+
+		DescribeTable(
+			"it returns the string representation of the type",
+			func(v dogma.Message, expected string) {
+				t := TypeOf(v)
+				Expect(t.String()).To(Equal(expected))
+			},
+			Entry("generic type", CommandA1, "stubs.CommandStub[TypeA]"),
+			Entry("pointer to generic type", &CommandA1, "*stubs.CommandStub[TypeA]"),
+			Entry("non-generic type", NonGenericCommand{}, "message_test.NonGenericCommand"),
+			Entry("pointer to non-generic type", &NonGenericCommand{}, "*message_test.NonGenericCommand"),
+		)
 	})
 })